day 9

+++ 2019/Day2.cs
@@ -9,16 +9,16 @@ namespace _2019
         [Fact]
         public void Test1()
         {
-            Assert.Equal(2, SolutionBeforePatch(new int[] { 1, 0, 0, 0, 99 }));
-            Assert.Equal(2, SolutionBeforePatch(new int[] { 2, 3, 0, 3, 99 }));
-            Assert.Equal(3500, SolutionBeforePatch(new int[] { 1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50 }));
-            Assert.Equal(30, SolutionBeforePatch(new int[] { 1, 1, 1, 4, 99, 5, 6, 0, 99 }));
+            Assert.Equal(2, SolutionBeforePatch(new long[] { 1, 0, 0, 0, 99 }));
+            Assert.Equal(2, SolutionBeforePatch(new long[] { 2, 3, 0, 3, 99 }));
+            Assert.Equal(3500, SolutionBeforePatch(new long[] { 1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50 }));
+            Assert.Equal(30, SolutionBeforePatch(new long[] { 1, 1, 1, 4, 99, 5, 6, 0, 99 }));
         }
 
         [Fact]
         public void Test2()
         {
-            var input = File.ReadAllText("input/day2.txt").Split(',').Select(int.Parse).ToArray();
+            var input = File.ReadAllText("input/day2.txt").Split(',').Select(long.Parse).ToArray();
             input[1] = 12;
             input[2] = 2;
             Assert.Equal(3790689, SolutionBeforePatch(input));
@@ -27,19 +27,19 @@ namespace _2019
         [Fact]
         public void Test3()
         {
-            var input = File.ReadAllText("input/day2.txt").Split(',').Select(int.Parse).ToArray();
+            var input = File.ReadAllText("input/day2.txt").Split(',').Select(long.Parse).ToArray();
 
             Assert.Equal(6533, Solution2(input));
         }
 
-        private int SolutionBeforePatch(int[] input)
+        private long SolutionBeforePatch(long[] input)
         {
             IntCodeEmulator emulator = new(input);
             emulator.Run();
             return emulator.ReadMemory(0);
         }
 
-        private int Solution2(int[] input)
+        private int Solution2(long[] input)
         {
             for (var i = 0; i <= 99; i++)
             {
